pyproject.toml
[tool.poetry]
name = "rag-tunning"
version = "0.1.0"
description = ""
authors = ["None"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.0"
openai = "^1.6.1"
weaviate-client = "^3.26.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.2"
faiss-cpu = "^1.7.4"
chromadb = "^0.4.22"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
bs4 = "^0.0.1"
termcolor = "^2.4.0"
msal = "^1.28.0"
pyjwt = "^2.8.0"
sseclient = "^0.0.27"
tempfileatexit = "^0.0.3"
langchain-openai = "^0.1.6"
duckduckgo-search = "^5.3.0"
langchainhub = "^0.1.15"
numexpr = "^2.10.0"
yfinance = "^0.2.40"
llama-index = "^0.10.48.post1"
llama-index-llms-openai = "^0.1.22"
llama-index-readers-json = "^0.1.5"
llama-index-llms-anthropic = "^0.1.14"
llama-index-vector-stores-chroma = "^0.1.9"
aiostream = "^0.6.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
